Python_Bridge/requirements.txt
pyserial
google-generativeai
SpeechRecognition
PyAudio
python-dotenv